Patch bytes preventing access to sequences below root
I have hit a roadblock when working using nanodicom. SequenceOfUltrasoundRegions 	(0018,6011) was unavailable to what turns out was parent_vr persisting from parsing, meaning that this tag was being read as byte.
Im unable to fix the root of the problem so here's 1 line that should fix most of its occurrences

THE BUG:
parent_vr being a global variable for a file means that after reading ('OB', 'OW', 'OX') data from a dicom file and not reading its contents will lead to some tags of type SQ not working.